Don't whack the rendezvous file if you are a tool

+++ src/mca/ptl/tcp/ptl_tcp_component.c
@@ -245,6 +245,7 @@ static int component_register(void)
 }
 
 static char *urifile = NULL;
+static bool created_rendezvous_file = false;
 
 static pmix_status_t component_open(void)
 {
@@ -312,7 +313,9 @@ pmix_status_t component_close(void)
         free(mca_ptl_tcp_component.pid_filename);
     }
     if (NULL != mca_ptl_tcp_component.rendezvous_filename) {
-        unlink(mca_ptl_tcp_component.rendezvous_filename);
+        if (created_rendezvous_file) {
+            unlink(mca_ptl_tcp_component.rendezvous_filename);
+        }
         free(mca_ptl_tcp_component.rendezvous_filename);
     }
     if (NULL != urifile) {
@@ -741,6 +744,7 @@ static pmix_status_t setup_listener(pmix_info_t info[], size_t ninfo,
             mca_ptl_tcp_component.rendezvous_filename = NULL;
             goto sockerror;
         }
+        created_rendezvous_file = true;
     }
 
   nextstep:
